Sign-compare errors

+++ core/test/dap_common_test.c
@@ -821,8 +821,8 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_sub(*(unsigned char*)l_a, *(unsigned int*)l_b) == dap_sub_builtin(*(unsigned char*)l_a, *(unsigned int*)l_b), "SUB UNSIGNED CHAR and UNSIGNED INT");
     dap_assert_PIF(dap_sub(*(unsigned short*)l_a, *(unsigned int*)l_b) == dap_sub_builtin(*(unsigned short*)l_a, *(unsigned int*)l_b), "SUB UNSIGNED SHORT and UNSIGNED INT");
     dap_assert_PIF(dap_sub(*(unsigned int*)l_a, *(unsigned int*)l_b) == dap_sub_builtin(*(unsigned int*)l_a, *(unsigned int*)l_b), "SUB UNSIGNED INT and UNSIGNED INT");
-    dap_assert_PIF(dap_sub(*(unsigned long*)l_a, *(unsigned int*)l_b) == dap_sub_builtin(*(unsigned long*)l_a, *(unsigned int*)l_b), "SUB UNSIGNED LONG and UNSIGNED INT");
-    dap_assert_PIF(dap_sub(*(unsigned long long*)l_a, *(unsigned int*)l_b) == dap_sub_builtin(*(unsigned long long*)l_a, *(unsigned int*)l_b), "SUB UNSIGNED LONG LONG and UNSIGNED INT");
+    dap_assert_PIF((unsigned long)dap_sub(*(unsigned long*)l_a, *(unsigned int*)l_b) == (unsigned long)dap_sub_builtin(*(unsigned long*)l_a, *(unsigned int*)l_b), "SUB UNSIGNED LONG and UNSIGNED INT");
+    dap_assert_PIF((unsigned long long)dap_sub(*(unsigned long long*)l_a, *(unsigned int*)l_b) == (unsigned long long)dap_sub_builtin(*(unsigned long long*)l_a, *(unsigned int*)l_b), "SUB UNSIGNED LONG LONG and UNSIGNED INT");
     // unsigned long
     dap_assert_PIF(dap_sub(*(char*)l_a, *(unsigned long*)l_b) == dap_sub_builtin(*(char*)l_a, *(unsigned long*)l_b), "SUB CHAR and UNSIGNED LONG");
     dap_assert_PIF(dap_sub(*(short*)l_a, *(unsigned long*)l_b) == dap_sub_builtin(*(short*)l_a, *(unsigned long*)l_b), "SUB SHORT and UNSIGNED LONG");
@@ -832,9 +832,9 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_sub(*(signed char*)l_a, *(unsigned long*)l_b) == dap_sub_builtin(*(signed char*)l_a, *(unsigned long*)l_b), "SUB SIGNED CHAR and UNSIGNED LONG");
     dap_assert_PIF(dap_sub(*(unsigned char*)l_a, *(unsigned long*)l_b) == dap_sub_builtin(*(unsigned char*)l_a, *(unsigned long*)l_b), "SUB UNSIGNED CHAR and UNSIGNED LONG");
     dap_assert_PIF(dap_sub(*(unsigned short*)l_a, *(unsigned long*)l_b) == dap_sub_builtin(*(unsigned short*)l_a, *(unsigned long*)l_b), "SUB UNSIGNED SHORT and UNSIGNED LONG");
-    dap_assert_PIF(dap_sub(*(unsigned int*)l_a, *(unsigned long*)l_b) == dap_sub_builtin(*(unsigned int*)l_a, *(unsigned long*)l_b), "SUB UNSIGNED INT and UNSIGNED LONG");
-    dap_assert_PIF(dap_sub(*(unsigned long*)l_a, *(unsigned long*)l_b) == dap_sub_builtin(*(unsigned long*)l_a, *(unsigned long*)l_b), "SUB UNSIGNED LONG and UNSIGNED LONG");
-    dap_assert_PIF(dap_sub(*(unsigned long long*)l_a, *(unsigned long*)l_b) == dap_sub_builtin(*(unsigned long long*)l_a, *(unsigned long*)l_b), "SUB UNSIGNED LONG LONG and UNSIGNED LONG");
+    dap_assert_PIF((unsigned int)dap_sub(*(unsigned int*)l_a, *(unsigned long*)l_b) == (unsigned int)dap_sub_builtin(*(unsigned int*)l_a, *(unsigned long*)l_b), "SUB UNSIGNED INT and UNSIGNED LONG");
+    dap_assert_PIF((unsigned long)dap_sub(*(unsigned long*)l_a, *(unsigned long*)l_b) == (unsigned long)dap_sub_builtin(*(unsigned long*)l_a, *(unsigned long*)l_b), "SUB UNSIGNED LONG and UNSIGNED LONG");
+    dap_assert_PIF((unsigned long long)dap_sub(*(unsigned long long*)l_a, *(unsigned long*)l_b) == (unsigned long long)dap_sub_builtin(*(unsigned long long*)l_a, *(unsigned long*)l_b), "SUB UNSIGNED LONG LONG and UNSIGNED LONG");
     // unsigned long long
     dap_assert_PIF(dap_sub(*(char*)l_a, *(unsigned long long*)l_b) == dap_sub_builtin(*(char*)l_a, *(unsigned long long*)l_b), "SUB CHAR and UNSIGNED LONG LONG");
     dap_assert_PIF(dap_sub(*(short*)l_a, *(unsigned long long*)l_b) == dap_sub_builtin(*(short*)l_a, *(unsigned long long*)l_b), "SUB SHORT and UNSIGNED LONG LONG");
@@ -844,8 +844,8 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_sub(*(signed char*)l_a, *(unsigned long long*)l_b) == dap_sub_builtin(*(signed char*)l_a, *(unsigned long long*)l_b), "SUB SIGNED CHAR and UNSIGNED LONG LONG");
     dap_assert_PIF(dap_sub(*(unsigned char*)l_a, *(unsigned long long*)l_b) == dap_sub_builtin(*(unsigned char*)l_a, *(unsigned long long*)l_b), "SUB UNSIGNED CHAR and UNSIGNED LONG LONG");
     dap_assert_PIF(dap_sub(*(unsigned short*)l_a, *(unsigned long long*)l_b) == dap_sub_builtin(*(unsigned short*)l_a, *(unsigned long long*)l_b), "SUB UNSIGNED SHORT and UNSIGNED LONG LONG");
-    dap_assert_PIF(dap_sub(*(unsigned int*)l_a, *(unsigned long long*)l_b) == dap_sub_builtin(*(unsigned int*)l_a, *(unsigned long long*)l_b), "SUB UNSIGNED INT and UNSIGNED LONG LONG");
-    dap_assert_PIF(dap_sub(*(unsigned long*)l_a, *(unsigned long long*)l_b) == dap_sub_builtin(*(unsigned long*)l_a, *(unsigned long long*)l_b), "SUB UNSIGNED LONG and UNSIGNED LONG LONG");
+    dap_assert_PIF((unsigned int)dap_sub(*(unsigned int*)l_a, *(unsigned long long*)l_b) == (unsigned int)dap_sub_builtin(*(unsigned int*)l_a, *(unsigned long long*)l_b), "SUB UNSIGNED INT and UNSIGNED LONG LONG");
+    dap_assert_PIF((unsigned long)dap_sub(*(unsigned long*)l_a, *(unsigned long long*)l_b) == (unsigned long)dap_sub_builtin(*(unsigned long*)l_a, *(unsigned long long*)l_b), "SUB UNSIGNED LONG and UNSIGNED LONG LONG");
     dap_assert_PIF((unsigned long long)dap_sub(*(unsigned long long*)l_a, *(unsigned long long*)l_b) == (unsigned long long)dap_sub_builtin(*(unsigned long long*)l_a, *(unsigned long long*)l_b), "SUB UNSIGNED LONG LONG and UNSIGNED LONG LONG");
 
 // MUL
